+++ include/boost/beast/websocket/detail/error.hpp
@@ -18,12 +18,18 @@ namespace boost {
 namespace beast {
 namespace websocket {
 enum class error;
+enum class condition;
 } // websocket
 } // beast
 
 namespace system {
 template<>
 struct is_error_code_enum<beast::websocket::error>
+{
+    static bool const value = true;
+};
+template<>
+struct is_error_condition_enum<beast::websocket::condition>
 {
     static bool const value = true;
 };
@@ -37,7 +43,7 @@ class error_codes : public error_category
 {
     template<class = void>
     string_view
-    get_message(error ev) const;
+    message(error e) const;
 
 public:
     const char*
@@ -49,7 +55,43 @@ public:
     std::string
     message(int ev) const override
     {
-        return get_message(static_cast<error>(ev)).to_string();
+        return message(
+            static_cast<error>(ev)).to_string();
+    }
+};
+
+class error_conditions : public error_category
+{
+    template<class = void>
+    string_view
+    message(condition c) const;
+
+    template<class = void>
+    bool
+    equivalent(error_code const& ec,
+        condition c) const noexcept;
+
+public:
+    const char*
+    name() const noexcept override
+    {
+        return "boost.beast.websocket";
+    }
+
+    std::string
+    message(int cv) const override
+    {
+        return message(
+            static_cast<condition>(cv)).to_string();
+    }
+
+    bool 
+    equivalent(
+        error_code const& ec,
+        int cv) const noexcept
+    {
+        return equivalent(ec,
+            static_cast<condition>(cv));
     }
 };
 
@@ -62,6 +104,15 @@ make_error_code(error e)
     static detail::error_codes const cat{};
     return error_code{static_cast<
         std::underlying_type<error>::type>(e), cat};
+}
+
+inline
+error_condition
+make_error_condition(condition c)
+{
+    static detail::error_conditions const cat{};
+    return error_condition{static_cast<
+        std::underlying_type<condition>::type>(c), cat};
 }
 
 } // websocket